# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "커피 키오스크 API",
    "description": "키오스크·백오피스·마이페이지 API. 인증 필요 엔드포인트는 Authorize에 JWT를 입력하세요.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "로컬 개발" }
  ],
...
  "paths": {
...
    "/api/admin/users": {
      "get": {
        "tags": ["Admin"],
        "summary": "회원 목록",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "성공", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/AdminUser" } } } } },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Admin only" }
        }
      }
    },
    "/api/admin/users/{id}": {
      "delete": {
        "tags": ["Admin"],
        "summary": "유저 삭제 (ADMIN 삭제 불가)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "204": { "description": "삭제 완료" },
          "403": { "description": "cannot delete admin" },
          "404": { "description": "user not found" }
        }
      }
    },
    "/api/admin/orders": {
      "get": {
        "tags": ["Admin"],
        "summary": "주문 목록",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["WAITING", "PREPARING", "PICKUP_READY", "COMPLETED", "CANCELED"] } }
        ],
        "responses": {
          "200": { "description": "성공" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Admin only" }
        }
      }
    },
    "/api/admin/orders/{id}": {
      "patch": {
        "tags": ["Admin"],
        "summary": "주문 상태 변경",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": { "status": { "type": "string", "enum": ["WAITING", "PREPARING", "PICKUP_READY", "COMPLETED", "CANCELED"] } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "성공" },
          "400": { "description": "invalid status" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Admin only" }
        }
      }
    },
    "/api/admin/products": {
      "get": {
        "tags": ["Admin"],
        "summary": "메뉴(상품) 목록",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "성공" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Admin only" }
        }
      },
      "post": {
        "tags": ["Admin"],
        "summary": "메뉴 등록",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["categoryId", "name", "basePrice"],
                "properties": {
                  "categoryId": { "type": "string" },
                  "name": { "type": "string" },
                  "englishName": { "type": "string" },
                  "description": { "type": "string" },
                  "basePrice": { "type": "number" },
                  "imageUrl": { "type": "string" },
                  "ingredients": { "type": "string" },
                  "calories": { "type": "string" },
                  "isAvailable": { "type": "boolean" },
                  "sortOrder": { "type": "number" },
                  "isBest": { "type": "boolean" },
                  "defaultShotCount": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "등록 성공" },
          "400": { "description": "categoryId, name, basePrice required" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Admin only" }
        }
      }
    },
    "/api/admin/products/{id}": {
      "patch": {
        "tags": ["Admin"],
        "summary": "메뉴 수정",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "categoryId": { "type": "string" },
                  "name": { "type": "string" },
                  "englishName": { "type": "string" },
                  "description": { "type": "string" },
                  "basePrice": { "type": "number" },
                  "imageUrl": { "type": "string" },
                  "ingredients": { "type": "string" },
                  "calories": { "type": "string" },
                  "isAvailable": { "type": "boolean" },
                  "sortOrder": { "type": "number" },
                  "isBest": { "type": "boolean" },
                  "defaultShotCount": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "성공" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Admin only" }
        }
      }
    },
    "/api/admin/terms": {
      "get": {
        "tags": ["Admin"],
        "summary": "약관 조회",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "성공", "content": { "application/json": { "schema": { "type": "object", "properties": { "content": { "type": "string" } } } } } },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Admin only" }
        }
      },
      "put": {
        "tags": ["Admin"],
        "summary": "약관 수정",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "content": { "type": "string" } } }
            }
          }
        },
        "responses": {
          "200": { "description": "성공" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Admin only" }
        }
      }
    },
    "/api/admin/privacy": {
      "get": {
        "tags": ["Admin"],
        "summary": "개인정보처리방침 조회",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "성공" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Admin only" }
        }
      },
      "put": {
        "tags": ["Admin"],
        "summary": "개인정보처리방침 수정",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "content": { "type": "string" } } }
            }
          }
        },
        "responses": {
          "200": { "description": "성공" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Admin only" }
        }
      }
    },
...
    "/api/orders": {
      "post": {
        "tags": ["Orders"],
        "summary": "주문 생성 (키오스크)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["totalPrice", "items"],
                "properties": {
                  "totalPrice": { "type": "number" },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": ["productId", "quantity"],
                      "properties": {
                        "productId": { "type": "string" },
                        "quantity": { "type": "number", "minimum": 1 },
                        "optionIds": { "type": "array", "items": { "type": "string" } }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "주문 생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "orderNumber": { "type": "string" },
                    "orderNo": { "type": "number" },
                    "orderId": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "유효성 오류 (INVALID_TOTAL_PRICE, ITEMS_REQUIRED 등)" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "로그인 응답의 token 값"
      }
    },
    "schemas": {
...
      "AdminUser": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "memberNo": { "type": "string" },
          "name": { "type": "string" },
          "username": { "type": "string" },
          "email": { "type": "string" },
          "orderCount": { "type": "number" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}